Add documentation using XML-comments

+++ ConsoleApp/Handlers/ContextMenu/AdminContextMenuHandler.cs
@@ -8,34 +8,60 @@ using System.Threading.Tasks;
 using StoreBLL.Interfaces;
 using StoreBLL.Models;
 
+/// <summary>
+/// Context menu handler for administrators.
+/// </summary>
 public class AdminContextMenuHandler : ContextMenuHandler
 {
+    /// <summary>
+    /// Initializes a new instance of the <see cref="AdminContextMenuHandler"/> class.
+    /// </summary>
+    /// <param name="service">The CRUD service.</param>
+    /// <param name="readModel">The function to read a model.</param>
     public AdminContextMenuHandler(ICrud service, Func<AbstractModel> readModel)
         : base(service, readModel)
     {
     }
 
+    /// <summary>
+    /// Adds an item.
+    /// </summary>
     public void AddItem()
     {
         this.service.Add(this.readModel());
     }
 
+    /// <summary>
+    /// Removes an item.
+    /// </summary>
     public void RemoveItem()
     {
         Console.WriteLine("Input record ID that will be removed");
-        int id = int.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
-        this.service.Delete(id);
+        var idInput = Console.ReadLine();
+        if (int.TryParse(idInput, out var id))
+        {
+            this.service.Delete(id);
+        }
     }
 
+    /// <summary>
+    /// Edits an item.
+    /// </summary>
     public void EditItem()
     {
         Console.WriteLine("Input record ID that will be edited");
-        int id = int.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
-        var record = this.readModel();
-        // TODO
-        this.service.Update(record);
+        var idInput = Console.ReadLine();
+        if (int.TryParse(idInput, out var id))
+        {
+            var record = this.readModel();
+            this.service.Update(record);
+        }
     }
 
+    /// <summary>
+    /// Generates the menu items.
+    /// </summary>
+    /// <returns>An array of menu items.</returns>
     public override (ConsoleKey id, string caption, Action action)[] GenerateMenuItems()
     {
         (ConsoleKey id, string caption, Action action)[] array =
